#  ddtrace/tracer: switch to tinylib/msgp for encoding/decoding