… ddtrace/vendor/github.com/philhofer/fwd/LICENSE.md
… ddtrace/vendor/github.com/philhofer/fwd/README.md
… ddtrace/vendor/github.com/philhofer/fwd/reader.go
… ddtrace/vendor/github.com/philhofer/fwd/reader_test.go
… ddtrace/vendor/github.com/philhofer/fwd/writer.go
… ddtrace/vendor/github.com/philhofer/fwd/writer_appengine.go
… ddtrace/vendor/github.com/philhofer/fwd/writer_test.go
… ddtrace/vendor/github.com/philhofer/fwd/writer_unsafe.go
… ddtrace/vendor/github.com/tinylib/msgp/.gitignore
… ddtrace/vendor/github.com/tinylib/msgp/.travis.yml
… ddtrace/vendor/github.com/tinylib/msgp/LICENSE
… ddtrace/vendor/github.com/tinylib/msgp/Makefile
… ddtrace/vendor/github.com/tinylib/msgp/README.md
… ddtrace/vendor/github.com/tinylib/msgp/_generated/convert.go
… ddtrace/vendor/github.com/tinylib/msgp/_generated/convert_test.go
… ddtrace/vendor/github.com/tinylib/msgp/_generated/def.go
… ddtrace/vendor/github.com/tinylib/msgp/_generated/def_test.go
… ddtrace/vendor/github.com/tinylib/msgp/_generated/gen_test.go
… ddtrace/vendor/github.com/tinylib/msgp/_generated/issue102.go
… ddtrace/vendor/github.com/tinylib/msgp/_generated/issue191.go
… ddtrace/vendor/github.com/tinylib/msgp/_generated/issue191_test.go
… ddtrace/vendor/github.com/tinylib/msgp/_generated/issue94.go
… ddtrace/vendor/github.com/tinylib/msgp/_generated/search.sh
… ddtrace/vendor/github.com/tinylib/msgp/gen/decode.go
… ddtrace/vendor/github.com/tinylib/msgp/gen/elem.go
… ddtrace/vendor/github.com/tinylib/msgp/gen/encode.go
… ddtrace/vendor/github.com/tinylib/msgp/gen/marshal.go
… ddtrace/vendor/github.com/tinylib/msgp/gen/size.go
… ddtrace/vendor/github.com/tinylib/msgp/gen/spec.go
… ddtrace/vendor/github.com/tinylib/msgp/gen/testgen.go