Benchmarks with `benchtime=1m`:

```
name                      old time/op    new time/op    delta
PayloadThroughput/10K-8     1.40ms ± 0%    0.69ms ± 0%  -50.69%
PayloadThroughput/100K-8    1.15ms ± 0%    0.63ms ± 0%  -45.79%
PayloadThroughput/1MB-8     1.20ms ± 0%    0.66ms ± 0%  -45.59%

name                      old alloc/op   new alloc/op   delta
PayloadThroughput/10K-8     46.5kB ± 0%    30.9kB ± 0%  -33.49%
PayloadThroughput/100K-8    4.93kB ± 0%    3.26kB ± 0%  -33.81%
PayloadThroughput/1MB-8       757B ± 0%      479B ± 0%  -36.72%

name                      old allocs/op  new allocs/op  delta
PayloadThroughput/10K-8      1.93k ± 0%     0.96k ± 0%  -50.05%
PayloadThroughput/100K-8       194 ± 0%        97 ± 0%  -50.00%
PayloadThroughput/1MB-8       20.0 ± 0%      10.0 ± 0%  -50.00%
```