@gbbr 2018-05-16
```
name                      old time/op    new time/op    delta
PayloadThroughput/10K-8     12.8µs ± 0%    15.8µs ± 0%   +23.36%
PayloadThroughput/100K-8     178µs ± 0%     175µs ± 0%    -1.65%
PayloadThroughput/1MB-8     1.49ms ± 0%    0.54ms ± 0%   -63.72%

name                      old speed      new speed      delta
PayloadThroughput/10K-8    808MB/s ± 0%   656MB/s ± 0%   -18.86%
PayloadThroughput/100K-8   581MB/s ± 0%   591MB/s ± 0%    +1.76%
PayloadThroughput/1MB-8    696MB/s ± 0%  1919MB/s ± 0%  +175.86%

name                      old alloc/op   new alloc/op   delta
PayloadThroughput/10K-8     28.5kB ± 0%    32.6kB ± 0%   +14.44%
PayloadThroughput/100K-8     278kB ± 0%     326kB ± 0%   +17.43%
PayloadThroughput/1MB-8     2.85MB ± 0%    3.26MB ± 0%   +14.72%

name                      old allocs/op  new allocs/op  delta
PayloadThroughput/10K-8       2.00 ± 0%      1.00 ± 0%   -50.00%
PayloadThroughput/100K-8      2.00 ± 0%      1.00 ± 0%   -50.00%
PayloadThroughput/1MB-8       2.00 ± 0%      1.00 ± 0%   -50.00%
```