@gbbr 2018-05-16
```
name                      old time/op    new time/op    delta
PayloadThroughput/10K-8     12.6µs ± 0%    16.1µs ± 0%  +27.54%
PayloadThroughput/100K-8     106µs ± 0%     173µs ± 0%  +63.02%
PayloadThroughput/1MB-8      885µs ± 0%     531µs ± 0%  -40.03%

name                      old speed      new speed      delta
PayloadThroughput/10K-8    822MB/s ± 0%   645MB/s ± 0%  -21.52%
PayloadThroughput/100K-8   976MB/s ± 0%   599MB/s ± 0%  -38.60%
PayloadThroughput/1MB-8   1.17GB/s ± 0%  1.95GB/s ± 0%  +66.89%

name                      old alloc/op   new alloc/op   delta
PayloadThroughput/10K-8     28.5kB ± 0%    32.6kB ± 0%  +14.44%
PayloadThroughput/100K-8     278kB ± 0%     326kB ± 0%  +17.43%
PayloadThroughput/1MB-8     2.85MB ± 0%    3.26MB ± 0%  +14.72%

name                      old allocs/op  new allocs/op  delta
PayloadThroughput/10K-8       2.00 ± 0%      1.00 ± 0%  -50.00%
PayloadThroughput/100K-8      2.00 ± 0%      1.00 ± 0%  -50.00%
PayloadThroughput/1MB-8       2.00 ± 0%      1.00 ± 0%  -50.00%
```