@gbbr 2018-05-16
```
gbbr@iMac:dd-trace-go.v1 (gbbr/tinylib)$ benchstat -delta-test none old-5s.txt new-5s.txt
name                      old time/op    new time/op     delta
PayloadThroughput/10K-8     1.44µs ± 0%     0.71µs ± 0%   -50.49%
PayloadThroughput/100K-8    12.3µs ± 0%      6.4µs ± 0%   -47.90%
PayloadThroughput/1MB-8      121µs ± 0%       63µs ± 0%   -47.85%

name                      old speed      new speed       delta
PayloadThroughput/10K-8   7.20GB/s ± 0%  14.54GB/s ± 0%  +102.05%
PayloadThroughput/100K-8  8.44GB/s ± 0%  16.22GB/s ± 0%   +92.13%
PayloadThroughput/1MB-8   8.55GB/s ± 0%  16.41GB/s ± 0%   +91.94%

name                      old alloc/op   new alloc/op    delta
PayloadThroughput/10K-8      53.0B ± 0%      35.0B ± 0%   -33.96%
PayloadThroughput/100K-8      102B ± 0%        63B ± 0%   -38.24%
PayloadThroughput/1MB-8       602B ± 0%       350B ± 0%   -41.86%

name                      old allocs/op  new allocs/op   delta
PayloadThroughput/10K-8       2.00 ± 0%       1.00 ± 0%   -50.00%
PayloadThroughput/100K-8      2.00 ± 0%       1.00 ± 0%   -50.00%
PayloadThroughput/1MB-8       2.00 ± 0%       1.00 ± 0%   -50.00%
```